+++ lunes_cms/static/js/toggle_plural_field.js
@@ -6,7 +6,16 @@ $(document).ready(() => {
   $("#id_word_type").change((event) =>
     $("#id_plural")
       .closest("div.row")
-      .toggle($(event.target).val() == "Nomen")
+      .toggle($(event.target).val() == "Nomen"),
+  );
+  $("#id_word_type").trigger("change");
+});
+
+$(document).ready(() => {
+  $("#id_word_type").change((event) =>
+    $("#id_grammatical_gender")
+      .closest("div.row")
+      .toggle($(event.target).val() == "Nomen"),
   );
   $("#id_word_type").trigger("change");
 });