+++ CHANGELOG.md
@@ -6,6 +6,7 @@ UNRELEASED
 * [ [#477](https://github.com/digitalfabrik/lunes-cms/issues/477) ] Add definition field
 * [ [#516](https://github.com/digitalfabrik/lunes-cms/issues/516) ] Add fields for plural of alternative words and plural article
 * [ [#352](https://github.com/digitalfabrik/lunes-cms/issues/352) ] Highlight unread feedback entries in list
+* [ [#436](https://github.com/digitalfabrik/lunes-cms/issues/436) ] Add fields for grammatical gender
 
 
 2023.7.0